src/python/shore/requirements.txt
celery==5.2.3
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-request-logging==0.7.3
djangorestframework==3.13.1
drf-yasg==1.20.0
environs==9.5.0
Faker==13.2.0
json2html==1.3.0
redis==4.1.4
requests==2.27.1
